Centrifuge snapshotter (interim)
