+++ package.json
@@ -2,14 +2,15 @@
   "name": "nedo",
   "version": "1.0.0",
   "description": "Create a snapshot of Interlay measures, e.g. vault collateral",
-  "main": "snapshot-interlay.js",
+  "main": "snapshot-centrifuge.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "dependencies": {
-    "@interlay/esplora-btc-api": "1.3.0",
-    "@interlay/interbtc-api": "^2.6.0",
-    "@interlay/monetary-js": "0.7.3",
+    "@centrifuge/centrifuge-js": "0.8.0",
+    "@polkadot/api-derive": "^10.12.2",
+    "rxjs": "^7.8.1",
+    "typescript": "^5.4.2",
     "@polkadot/api": "10.9.1",
     "big.js": "6.1.1",
     "bitcoinjs-lib": "^5.2.0",
